Chapter 1 - Tuples, Points, and Vectors
Provide implementation for foundation classes that will be used as
the basis for subsequent functionality.  Created a 4-tuple class,
Tuple, and then derived Point and Vector classes from that.

Provide support for the following operations on tuples:
* checking for equality/inequality
* addition/subtraction
* negation
* scalar multiplication/division

Provide additional support for the following operations on vectors
only:
* magnitude
* normalization
* dot product
* cross product

Add appropriate unit tests.