Refactor Tuple and Vector
Some decisions made early on about adding static and instance
methods on the Vector class turn out to make things a little more
cumbersome.  There are a couple of ways to handle this, but in
order to make progress in the book, move these methods to the
Tuple class and simply check for vectors to make sure that they
are not called with points.